Tidied up labels in GUI for SuperQ
Fixed operational test for GLPK, this didn't always work before.

Explicitly added colt as a dependency for joptimizer.  However, I still need to come up with a script to download and install to maven joptimizer itself and whatever other dependencies are required... unfortunately, it looks like the installation procedure for joptimizer can't be fully automated with maven as they are not hosting the joptimizer jar on a remote maven repo! :(