We keep quartets in canonicalised form throughout quart (was chopper) and only convert to grouped quartet form at the last step.  This makes it easier to debug what's going on and we've increased performance significantly as a by product.
Quart has a proper integration test now which combines two quartet systems together, that rpoduce a conflict.  All values are correct as far as I can tell.  We should put more tests in however but we'll do this as we go.

SuperQ and QNet are running but I think we still have a bit of work todo before they work correctly.