Moved network drawing code into core... just did a mass copy and paste here... nothing fancy.
Checked that the viewer works.  It seems to work ok for flat networks.  Still need to test with circular split networks but this should be fine.

Modified nexus reader so that we can handle '/' characters in identifiers.

Seems to be a problem in FlatNJ where it gets stuck in the collectAndSortBoxesForTheSplit method in BoxOpener... need to investigate.

Also need to test Andreas' new code to see if we can create a network from the circular split system.